+++ pkg/backend/backend.go
@@ -150,7 +150,9 @@ type Backend interface {
 	// ListStacks returns a list of stack summaries for all known stacks in the target backend.
 	ListStacks(ctx context.Context, filter ListStacksFilter) ([]StackSummary, error)
 
-	RenameStack(ctx context.Context, stack Stack, newName tokens.QName) error
+	// RenameStack renames the given stack to a new name, and then returns an updated stack reference that
+	// can be used to refer to the newly renamed stack.
+	RenameStack(ctx context.Context, stack Stack, newName tokens.QName) (StackReference, error)
 
 	// Preview shows what would be updated given the current workspace's contents.
 	Preview(ctx context.Context, stack Stack, op UpdateOperation) (engine.ResourceChanges, result.Result)